Manually resolves links relatively to `root_dir`, and prevent escape
This patch is a followup of #311 (2a89f76c).

It appeared that we were not resolving paths when reading from files.
This means that a symbolic link present under `root_dir` could be
blindly followed _outside_ of `root_dir`, possibly leading to host
files.

Note : this patch **only** changes `root_dir` behavior.